// generated_library/src/types/Types.ts
export type BillingRequest = {
  
    // List of actions that can be performed before this billing request can be
  // fulfilled.
  actions?: BillingRequestAction[]

  // Fixed [timestamp](#api-usage-time-zones--dates), recording when this
  // resource was created.
  created_at?: string

  // (Optional) If true, this billing request can fallback from instant payment
  // to direct debit.
  // Should not be set if GoCardless payment intelligence feature is used.
  // 
  // See [Billing Requests: Retain customers with
  // Fallbacks](https://developer.gocardless.com/billing-requests/retain-customers-with-fallbacks/)
  // for more information.
  fallback_enabled?: boolean

  // True if the billing request was completed with direct debit.
  fallback_occurred?: boolean

  // Unique identifier, beginning with "BRQ".
  id: string

  // Request for an instalment schedule. Has to contain either
  // `instalments_with_schedule` object or an array of `instalments_with_dates`
  // objects
  instalment_schedule_request?: BillingRequestInstalmentScheduleRequest|null


  // Resources linked to this BillingRequest.
  links?: BillingRequestLinks

  // Request for a mandate
  mandate_request?: BillingRequestMandateRequest

  // Key-value store of custom data. Up to 3 keys are permitted, with key names
  // up to 50 characters and values up to 500 characters.
  metadata?: JsonMap

  // Request for a one-off strongly authorised payment
  payment_request?: BillingRequestPaymentRequest

  // Specifies the high-level purpose of a mandate and/or payment using a set of
  // pre-defined categories. Required for the PayTo scheme, optional for all
  // others. Currently `mortgage`, `utility`, `loan`, `dependant_support`,
  // `gambling`, `retail`, `salary`, `personal`, `government`, `pension`, `tax`
  // and `other` are supported.
  purpose_code?: BillingRequestPurposeCode

  // 
  resources?: BillingRequestResources

  // One of:
  // <ul>
  // <li>`pending`: the billing request is pending and can be used</li>
  // <li>`ready_to_fulfil`: the billing request is ready to fulfil</li>
  // <li>`fulfilling`: the billing request is currently undergoing
  // fulfilment</li>
  // <li>`fulfilled`: the billing request has been fulfilled and a payment
  // created</li>
  // <li>`cancelled`: the billing request has been cancelled and cannot be
  // used</li>
  // </ul>
  status?: BillingRequestStatus

  // Request for a subscription
  subscription_request?: BillingRequestSubscriptionRequest|null

}




  
    
      
      
        
          
          
            
          
        
      
        
        
      
        
            

            
/** Type for a billingrequestcreaterequestlinks resource. */
export type BillingRequestCreateRequestLinks = {
  
    // ID of the associated [creditor](#core-endpoints-creditors). Only required
  // if your account manages multiple creditors.
  creditor?: string

  // ID of the [customer](#core-endpoints-customers) against which this request
  // should be made.
  customer?: string

  // (Optional) ID of the
  // [customer_bank_account](#core-endpoints-customer-bank-accounts) against
  // which this request should be made.
  // 
  customer_bank_account?: string

}
// ...
export type BillingRequestCustomer = {
  
    // Customer's company name. Required unless a `given_name` and `family_name`
  // are provided. For Canadian customers, the use of a `company_name` value
  // will mean that any mandate created from this customer will be considered to
  // be a "Business PAD" (otherwise, any mandate will be considered to be a
  // "Personal PAD").
  company_name?: string|null

  // Customer's email address. Required in most cases, as this allows GoCardless
  // to send notifications to this customer.
  email?: string|null

  // Customer's surname. Required unless a `company_name` is provided.
  family_name?: string|null

  // Customer's first name. Required unless a `company_name` is provided.
  given_name?: string|null

  //  [ISO 639-1](http://en.wikipedia.org/wiki/List_of_ISO_639-1_codes) code.
  // Used as the language for notification emails sent by GoCardless if your
  // organisation does not send its own (see [compliance
  // requirements](#appendix-compliance-requirements)). Currently only "en",
  // "fr", "de", "pt", "es", "it", "nl", "da", "nb", "sl", "sv" are supported.
  // If this is not provided and a customer was linked during billing request
  // creation, the linked customer language will be used. Otherwise, the
  // language is default to "en".
  language?: string|null

  // Key-value store of custom data. Up to 3 keys are permitted, with key names
  // up to 50 characters and values up to 500 characters.
  metadata?: JsonMap

  // [ITU E.123](https://en.wikipedia.org/wiki/E.123) formatted phone number,
  // including country code.
  phone_number?: string|null

}
// ...
                export enum BillingRequestAccountType {
  Savings = "savings",
  Checking = "checking",
}
// ...
export type BillingRequestInstalmentScheduleRequest = {
  
    // The amount to be deducted from each payment as an app fee, to be paid to
  // the partner integration which created the subscription, in the lowest
  // denomination for the currency (e.g. pence in GBP, cents in EUR).
  app_fee?: string|null

  // [ISO 4217](http://en.wikipedia.org/wiki/ISO_4217#Active_codes) currency
  // code. Currently "USD" and "CAD" are supported.
  currency?: string

  // An explicit array of instalment payments, each specifying at least an
  // `amount` and `charge_date`. See [create (with
  // dates)](#instalment-schedules-create-with-dates)
  instalments_with_dates?: BillingRequestInstalmentScheduleRequestInstalmentsWithDate[]|null

  // Frequency of the payments you want to create, together with an array of
  // payment
  // amounts to be collected, with a specified start date for the first payment.
  // See [create (with schedule)](#instalment-schedules-create-with-schedule)
  // 
  instalments_with_schedule?: BillingRequestInstalmentScheduleRequestInstalmentsWithSchedule|null


  // Resources linked to this BillingRequestInstalmentScheduleRequest.
  links?: BillingRequestInstalmentScheduleRequestLinks

  // Key-value store of custom data. Up to 3 keys are permitted, with key names
  // up to 50 characters and values up to 500 characters.
  metadata?: JsonMap

  // Name of the instalment schedule, up to 100 chars. This name will also be
  // copied to the payments of the instalment schedule if you use schedule-based
  // creation.
  name?: string

  // An optional payment reference. This will be set as the reference on each
  // payment
  // created and will appear on your customer's bank statement. See the
  // documentation for
  // the [create payment endpoint](#payments-create-a-payment) for more details.
  // <br />
  payment_reference?: string|null

  // On failure, automatically retry payments using [intelligent
  // retries](#success-intelligent-retries). Default is `false`. <p
  // class="notice"><strong>Important</strong>: To be able to use intelligent
  // retries, Success+ needs to be enabled in [GoCardless
  // dashboard](https://manage.gocardless.com/success-plus). </p>
  retry_if_possible?: boolean

  // The total amount of the instalment schedule, defined as the sum of all
  // individual
  // payments, in the lowest denomination for the currency (e.g. pence in GBP,
  // cents in
  // EUR). If the requested payment amounts do not sum up correctly, a
  // validation error
  // will be returned.
  total_amount?: string

}
// ...
export type BillingRequestMandateRequest = {
  
    // This field is ACH specific, sometimes referred to as [SEC
  // code](https://www.moderntreasury.com/learn/sec-codes).
  // 
  // This is the way that the payer gives authorisation to the merchant.
  //   web: Authorisation is Internet Initiated or via Mobile Entry (maps to SEC
  // code: WEB)
  //   telephone: Authorisation is provided orally over telephone (maps to SEC
  // code: TEL)
  //   paper: Authorisation is provided in writing and signed, or similarly
  // authenticated (maps to SEC code: PPD)
  // 
  authorisation_source?: BillingRequestMandateRequestAuthorisationSource

  // This attribute represents the authorisation type between the payer and
  // merchant. It can be set to `one_off`,
  // `recurring` or `standing` for ACH scheme. And `single`, `recurring` and
  // `sporadic` for PAD scheme. _Note:_ This is only supported for ACH and PAD
  // schemes.
  // 
  consent_type?: string|null

  // Constraints that will apply to the mandate_request. (Optional) Specifically
  // for PayTo and VRP.
  constraints?: BillingRequestMandateRequestConstraints|null

  // [ISO 4217](http://en.wikipedia.org/wiki/ISO_4217#Active_codes) currency
  // code.
  currency?: string

  // A human-readable description of the payment and/or mandate. This will be
  // displayed to the payer when authorising the billing request.
  // 
  description?: string|null


  // Resources linked to this BillingRequestMandateRequest.
  links?: BillingRequestMandateRequestLinks

  // Key-value store of custom data. Up to 3 keys are permitted, with key names
  // up to 50 characters and values up to 500 characters.
  metadata?: JsonMap

  // This attribute can be set to true if the payer has indicated that multiple
  // signatures are required for the mandate. As long as every other Billing
  // Request actions have been completed, the payer will receive an email
  // notification containing instructions on how to complete the additional
  // signature. The dual signature flow can only be completed using GoCardless
  // branded pages.
  payer_requested_dual_signature?: boolean

  // A bank payment scheme. Currently "ach", "autogiro", "bacs", "becs",
  // "becs_nz", "betalingsservice", "faster_payments", "pad", "pay_to" and
  // "sepa_core" are supported. Optional for mandate only requests - if left
  // blank, the payer will be able to select the currency/scheme to pay with
  // from a list of your available schemes.
  scheme?: string|null

  // If true, this billing request would be used to set up a mandate solely for
  // moving (or sweeping) money from one account owned by the payer to another
  // account that the payer also owns. This is required for Faster Payments
  sweeping?: boolean

  // Verification preference for the mandate. One of:
  // <ul>
  //   <li>`minimum`: only verify if absolutely required, such as when part of
  // scheme rules</li>
  //   <li>`recommended`: in addition to `minimum`, use the GoCardless payment
  // intelligence solution to decide if a payer should be verified</li>
  //   <li>`when_available`: if verification mechanisms are available, use
  // them</li>
  //   <li>`always`: as `when_available`, but fail to create the Billing Request
  // if a mechanism isn't available</li>
  // </ul>
  // 
  // By default, all Billing Requests use the `recommended` verification
  // preference. It uses GoCardless payment intelligence solution to determine
  // if a payer is fraudulent or not. The verification mechanism is based on the
  // response and the payer may be asked to verify themselves. If the feature is
  // not available, `recommended` behaves like `minimum`.
  // 
  // If you never wish to take advantage of our reduced risk products and
  // Verified Mandates as they are released in new schemes, please use the
  // `minimum` verification preference.
  // 
  // See [Billing Requests: Creating Verified
  // Mandates](https://developer.gocardless.com/getting-started/billing-requests/verified-mandates/)
  // for more information.
  verify?: BillingRequestMandateRequestVerify

}
// ...
    export enum BillingRequestPurposeCode {
  Mortgage = "mortgage",
  Utility = "utility",
  Loan = "loan",
  DependantSupport = "dependant_support",
  Gambling = "gambling",
  Retail = "retail",
  Salary = "salary",
  Personal = "personal",
  Government = "government",
  Pension = "pension",
  Tax = "tax",
  Other = "other",
}
// ...
export interface JsonMap { [key: string]: JsonField | JsonMap | JsonArray; }
export interface JsonArray extends Array<JsonField> {}

export type APIResponse = {
  __response__: object,
}
// ...
export interface ListMeta {
  limit: number
  cursors: ListMetaCursor
}